requirements.txt
acres==0.2.0
annotated-types==0.7.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
cryptography==44.0.0
cssselect2==0.7.0
defusedxml==0.7.1
Django==5.1.4
django-allauth==65.14.0
django-cors-headers==4.6.0
django-filter==24.3
django-nepali-datetime-field==0.8.0
django-qr-code==4.1.0
django-tinymce==4.1.0
django-unfold==0.43.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
etelemetry==0.3.1
filelock==3.16.1
gunicorn==23.0.0
html5lib==1.1
httplib2==0.22.0
idna==3.10
isodate==0.6.1
looseversion==1.3.0
lxml==5.3.0
nepali-datetime==1.0.8.3
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.1
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==12.1.0
prov==2.0.1
psycopg2-binary==2.9.11
puremagic==1.28
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
pydot==4.0.1
pyHanko==0.28.0
pyhanko-certvalidator==0.27.0
PyJWT==2.10.1
PyMuPDF==1.25.1
pyparsing==3.3.1
pypdf==5.1.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
qrcode==8.0
rdflib==6.3.2
reportlab==4.2.5
requests==2.32.3
scipy==1.16.3
segno==1.6.6
simplejson==3.19.3
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.4.0
traits==6.4.3
typing-inspection==0.4.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.3.0
webencodings==0.5.1
whitenoise==6.8.2
xhtml2pdf==0.2.16
